Scene Loader Cleanup (#7841)
* remove PreventBakeTagComponent

* eliminate defaultData and some more de/serializers

* cleanup log

* refactor loop animation component loader

* remove envmap deserializer

* remove unnecessary setcomponent helper functions
