+++ packages/engine/src/physics/systems/PhysicsSystem.ts
@@ -115,9 +115,7 @@ export function smoothVelocityBasedKinematicBody(entity: Entity, dt: number, sub
 
 export default async function PhysicsSystem() {
   Engine.instance.sceneComponentRegistry.set(ColliderComponent.name, SCENE_COMPONENT_COLLIDER)
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_COLLIDER, {
-    defaultData: {}
-  })
+  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_COLLIDER, {})
 
   Engine.instance.scenePrefabRegistry.set(PhysicsPrefabs.collider, [
     { name: SCENE_COMPONENT_TRANSFORM, props: SCENE_COMPONENT_TRANSFORM_DEFAULT_VALUES },
