+++ packages/engine/src/scene/components/EnvmapComponent.ts
@@ -15,16 +15,15 @@ import {
   Vector3
 } from 'three'
 
-import { getMutableState, useHookstate } from '@etherealengine/hyperflux'
+import { getMutableState, getState, useHookstate } from '@etherealengine/hyperflux'
 
 import { AssetLoader } from '../../assets/classes/AssetLoader'
 import { isClient } from '../../common/functions/isClient'
-import { Engine } from '../../ecs/classes/Engine'
 import { SceneState } from '../../ecs/classes/Scene'
 import {
   defineComponent,
-  getMutableComponent,
   getOptionalComponent,
+  removeComponent,
   useComponent,
   useOptionalComponent
 } from '../../ecs/functions/ComponentFunctions'
@@ -35,9 +34,7 @@ import { addError, removeError } from '../functions/ErrorFunctions'
 import { EnvMapBakeTypes } from '../types/EnvMapBakeTypes'
 import { EnvMapBakeComponent } from './EnvMapBakeComponent'
 import { GroupComponent } from './GroupComponent'
-import { ModelComponent } from './ModelComponent'
 
-const tempVector = new Vector3()
 const tempColor = new Color()
 
 export const EnvmapComponent = defineComponent({
@@ -53,6 +50,7 @@ export const EnvmapComponent = defineComponent({
   },
 
   onSet: (entity, component, json) => {
+    if (entity === getState(SceneState).sceneEntity) removeComponent(entity, EnvmapComponent)
     if (typeof json?.type === 'string') component.type.set(json.type)
     if (typeof json?.envMapTextureType === 'string') component.envMapTextureType.set(json.envMapTextureType)
     if (typeof json?.envMapSourceColor === 'number') component.envMapSourceColor.set(new Color(json.envMapSourceColor))
