+++ packages/engine/src/scene/systems/EnvironmentSystem.ts
@@ -13,7 +13,6 @@ import { EnvMapBakeComponent, SCENE_COMPONENT_ENVMAP_BAKE } from '../components/
 import { EnvmapComponent, SCENE_COMPONENT_ENVMAP } from '../components/EnvmapComponent'
 import { SCENE_COMPONENT_SKYBOX, SkyboxComponent } from '../components/SkyboxComponent'
 import { SCENE_COMPONENT_VISIBLE } from '../components/VisibleComponent'
-import { deserializeEnvMap } from '../functions/loaders/EnvMapFunctions'
 import { ScenePrefabs } from './SceneObjectUpdateSystem'
 
 export default async function EnvironmentSystem() {
@@ -23,9 +22,7 @@ export default async function EnvironmentSystem() {
   ])
 
   Engine.instance.sceneComponentRegistry.set(SkyboxComponent.name, SCENE_COMPONENT_SKYBOX)
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_SKYBOX, {
-    defaultData: {}
-  })
+  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_SKYBOX, {})
 
   Engine.instance.scenePrefabRegistry.set(ScenePrefabs.envMapbake, [
     { name: SCENE_COMPONENT_TRANSFORM, props: SCENE_COMPONENT_TRANSFORM_DEFAULT_VALUES },
@@ -34,14 +31,10 @@ export default async function EnvironmentSystem() {
   ])
 
   Engine.instance.sceneComponentRegistry.set(EnvMapBakeComponent.name, SCENE_COMPONENT_ENVMAP_BAKE)
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_ENVMAP_BAKE, {
-    defaultData: {}
-  })
+  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_ENVMAP_BAKE, {})
 
   Engine.instance.sceneComponentRegistry.set(EnvmapComponent.name, SCENE_COMPONENT_ENVMAP)
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_ENVMAP, {
-    deserialize: deserializeEnvMap
-  })
+  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_ENVMAP, {})
 
   const reactor = startReactor(() => {
     const background = useHookstate(getMutableState(SceneState).background)
