+++ packages/engine/src/scene/systems/LoadVolumeSystem.ts
@@ -6,9 +6,7 @@ export default async function LoadVolumeSystem() {
   Engine.instance.sceneComponentRegistry.set(LoadVolumeComponent.name, SCENE_COMPONENT_LOAD_VOLUME)
   Engine.instance.scenePrefabRegistry.set(ScenePrefabs.loadVolume, [{ name: SCENE_COMPONENT_LOAD_VOLUME, props: {} }])
 
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_LOAD_VOLUME, {
-    defaultData: { targets: {} }
-  })
+  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_LOAD_VOLUME, {})
 
   return {}
 }
