+++ packages/engine/src/scene/systems/SceneObjectUpdateSystem.ts
@@ -1,5 +1,3 @@
-import { Not } from 'bitecs'
-
 import { ComponentJson } from '@etherealengine/common/src/interfaces/SceneInterface'
 import { createActionQueue, removeActionQueue } from '@etherealengine/hyperflux'
 
@@ -33,8 +31,6 @@ import {
 } from '../components/OceanComponent'
 import { ParticleSystemComponent, SCENE_COMPONENT_PARTICLE_SYSTEM } from '../components/ParticleSystemComponent'
 import { PrefabComponent, SCENE_COMPONENT_PREFAB } from '../components/PrefabComponent'
-import { PreventBakeTagComponent, SCENE_COMPONENT_PREVENT_BAKE } from '../components/PreventBakeTagComponent'
-import { SceneAssetPendingTagComponent } from '../components/SceneAssetPendingTagComponent'
 import { SCENE_COMPONENT_SCENE_PREVIEW_CAMERA, ScenePreviewCameraComponent } from '../components/ScenePreviewCamera'
 import { SCENE_COMPONENT_SCREENSHARETARGET, ScreenshareTargetComponent } from '../components/ScreenshareTargetComponent'
 import {
@@ -48,9 +44,7 @@ import { SCENE_COMPONENT_SYSTEM, SystemComponent } from '../components/SystemCom
 import { SCENE_COMPONENT_VISIBLE, VisibleComponent } from '../components/VisibleComponent'
 import { SCENE_COMPONENT_WATER, WaterComponent } from '../components/WaterComponent'
 import { deserializeCloud, serializeCloud, updateCloud } from '../functions/loaders/CloudFunctions'
-import { deserializeGroup } from '../functions/loaders/GroupFunctions'
 import { deserializeInterior, serializeInterior, updateInterior } from '../functions/loaders/InteriorFunctions'
-import { serializeLoopAnimation, updateLoopAnimation } from '../functions/loaders/LoopAnimationFunctions'
 import { deserializeModel } from '../functions/loaders/ModelFunctions'
 import { deserializeOcean, serializeOcean, updateOcean } from '../functions/loaders/OceanFunctions'
 import { deserializePrefab } from '../functions/loaders/PrefabComponentFunctions'
@@ -95,12 +89,7 @@ export default async function SceneObjectUpdateSystem() {
   Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_VISIBLE, {})
 
   Engine.instance.sceneComponentRegistry.set(ShadowComponent.name, SCENE_COMPONENT_SHADOW)
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_SHADOW, {
-    defaultData: true
-  })
-
-  Engine.instance.sceneComponentRegistry.set(PreventBakeTagComponent.name, SCENE_COMPONENT_PREVENT_BAKE)
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_PREVENT_BAKE, {})
+  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_SHADOW, {})
 
   /**
    * Metadata
@@ -118,9 +107,7 @@ export default async function SceneObjectUpdateSystem() {
   Engine.instance.scenePrefabRegistry.set(ScenePrefabs.system, [{ name: SCENE_COMPONENT_SYSTEM, props: {} }])
 
   Engine.instance.sceneComponentRegistry.set(SystemComponent.name, SCENE_COMPONENT_SYSTEM)
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_SYSTEM, {
-    defaultData: {}
-  })
+  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_SYSTEM, {})
 
   Engine.instance.scenePrefabRegistry.set(ScenePrefabs.spawnPoint, [
     { name: SCENE_COMPONENT_TRANSFORM, props: SCENE_COMPONENT_TRANSFORM_DEFAULT_VALUES },
@@ -129,9 +116,7 @@ export default async function SceneObjectUpdateSystem() {
   ])
 
   Engine.instance.sceneComponentRegistry.set(SpawnPointComponent.name, SCENE_COMPONENT_SPAWN_POINT)
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_SPAWN_POINT, {
-    defaultData: {}
-  })
+  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_SPAWN_POINT, {})
 
   /**
    * Assets
@@ -174,10 +159,7 @@ export default async function SceneObjectUpdateSystem() {
   ])
 
   Engine.instance.sceneComponentRegistry.set(GroupComponent.name, SCENE_COMPONENT_GROUP)
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_GROUP, {
-    deserialize: deserializeGroup,
-    serialize: () => undefined!
-  })
+  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_GROUP, {})
 
   Engine.instance.scenePrefabRegistry.set(ScenePrefabs.groundPlane, [
     { name: SCENE_COMPONENT_TRANSFORM, props: SCENE_COMPONENT_TRANSFORM_DEFAULT_VALUES },
@@ -187,9 +169,7 @@ export default async function SceneObjectUpdateSystem() {
   ])
 
   Engine.instance.sceneComponentRegistry.set(GroundPlaneComponent.name, SCENE_COMPONENT_GROUND_PLANE)
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_GROUND_PLANE, {
-    defaultData: {}
-  })
+  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_GROUND_PLANE, {})
 
   Engine.instance.scenePrefabRegistry.set(ScenePrefabs.image, [
     ...defaultSpatialComponents,
@@ -200,15 +180,10 @@ export default async function SceneObjectUpdateSystem() {
   ])
 
   Engine.instance.sceneComponentRegistry.set(ImageComponent.name, SCENE_COMPONENT_IMAGE)
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_IMAGE, {
-    defaultData: {}
-  })
+  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_IMAGE, {})
 
   Engine.instance.sceneComponentRegistry.set(LoopAnimationComponent.name, SCENE_COMPONENT_LOOP_ANIMATION)
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_LOOP_ANIMATION, {
-    defaultData: {},
-    serialize: serializeLoopAnimation
-  })
+  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_LOOP_ANIMATION, {})
 
   Engine.instance.scenePrefabRegistry.set(ScenePrefabs.cloud, [
     ...defaultSpatialComponents,
@@ -217,7 +192,6 @@ export default async function SceneObjectUpdateSystem() {
 
   Engine.instance.sceneComponentRegistry.set(CloudComponent.name, SCENE_COMPONENT_CLOUD)
   Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_CLOUD, {
-    defaultData: SCENE_COMPONENT_CLOUD_DEFAULT_VALUES,
     deserialize: deserializeCloud,
     serialize: serializeCloud
   })
@@ -228,7 +202,6 @@ export default async function SceneObjectUpdateSystem() {
 
   Engine.instance.sceneComponentRegistry.set(OceanComponent.name, SCENE_COMPONENT_OCEAN)
   Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_OCEAN, {
-    defaultData: SCENE_COMPONENT_OCEAN_DEFAULT_VALUES,
     deserialize: deserializeOcean,
     serialize: serializeOcean
   })
@@ -238,9 +211,7 @@ export default async function SceneObjectUpdateSystem() {
   ])
 
   Engine.instance.sceneComponentRegistry.set(WaterComponent.name, SCENE_COMPONENT_WATER)
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_WATER, {
-    defaultData: {}
-  })
+  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_WATER, {})
 
   Engine.instance.scenePrefabRegistry.set(ScenePrefabs.interior, [
     ...defaultSpatialComponents,
@@ -249,7 +220,6 @@ export default async function SceneObjectUpdateSystem() {
 
   Engine.instance.sceneComponentRegistry.set(InteriorComponent.name, SCENE_COMPONENT_INTERIOR)
   Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_INTERIOR, {
-    defaultData: SCENE_COMPONENT_INTERIOR_DEFAULT_VALUES,
     deserialize: deserializeInterior,
     serialize: serializeInterior
   })
@@ -261,7 +231,6 @@ export default async function SceneObjectUpdateSystem() {
 
   Engine.instance.sceneComponentRegistry.set(SplineComponent.name, SCENE_COMPONENT_SPLINE)
   Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_SPLINE, {
-    defaultData: {},
     deserialize: deserializeSpline,
     serialize: serializeSpline
   })
@@ -273,11 +242,8 @@ export default async function SceneObjectUpdateSystem() {
 
   Engine.instance.sceneComponentRegistry.set(ParticleSystemComponent.name, SCENE_COMPONENT_PARTICLE_SYSTEM)
 
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_PARTICLE_SYSTEM, {
-    defaultData: {}
-  })
+  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_PARTICLE_SYSTEM, {})
 
-  const loopableAnimationQuery = defineQuery([LoopAnimationComponent, Not(SceneAssetPendingTagComponent)])
   const cloudQuery = defineQuery([CloudComponent])
   const oceanQuery = defineQuery([OceanComponent])
   const interiorQuery = defineQuery([InteriorComponent])
@@ -288,14 +254,12 @@ export default async function SceneObjectUpdateSystem() {
   const execute = () => {
     for (const action of modifyPropertyActionQueue()) {
       for (const entity of action.entities) {
-        if (hasComponent(entity, LoopAnimationComponent)) updateLoopAnimation(entity)
         if (hasComponent(entity, CloudComponent)) updateCloud(entity)
         if (hasComponent(entity, OceanComponent)) updateOcean(entity)
         if (hasComponent(entity, InteriorComponent)) updateInterior(entity)
       }
     }
 
-    for (const entity of loopableAnimationQuery.enter()) updateLoopAnimation(entity)
     for (const entity of cloudQuery.enter()) updateCloud(entity)
     for (const entity of oceanQuery.enter()) updateOcean(entity)
     for (const entity of interiorQuery.enter()) updateInterior(entity)
@@ -308,9 +272,6 @@ export default async function SceneObjectUpdateSystem() {
 
     Engine.instance.sceneComponentRegistry.delete(ShadowComponent.name)
     Engine.instance.sceneLoadingRegistry.delete(SCENE_COMPONENT_SHADOW)
-
-    Engine.instance.sceneComponentRegistry.delete(PreventBakeTagComponent.name)
-    Engine.instance.sceneLoadingRegistry.delete(SCENE_COMPONENT_PREVENT_BAKE)
 
     /**
      * Metadata
@@ -397,7 +358,6 @@ export default async function SceneObjectUpdateSystem() {
     Engine.instance.sceneComponentRegistry.delete(ParticleSystemComponent.name)
     Engine.instance.sceneLoadingRegistry.delete(SCENE_COMPONENT_PARTICLE_SYSTEM)
 
-    removeQuery(loopableAnimationQuery)
     removeQuery(cloudQuery)
     removeQuery(oceanQuery)
     removeQuery(interiorQuery)
