+++ packages/engine/src/transform/systems/TransformSystem.ts
@@ -26,7 +26,6 @@ import {
   RigidBodyKinematicVelocityBasedTagComponent
 } from '../../physics/components/RigidBodyComponent'
 import { GroupComponent } from '../../scene/components/GroupComponent'
-import { deserializeTransform, serializeTransform } from '../../scene/functions/loaders/TransformFunctions'
 import { ComputedTransformComponent } from '../components/ComputedTransformComponent'
 import {
   DistanceFromCameraComponent,
@@ -164,11 +163,7 @@ const getDistanceSquaredFromTarget = (entity: Entity, targetPosition: Vector3) =
 
 export default async function TransformSystem() {
   Engine.instance.sceneComponentRegistry.set(TransformComponent.name, SCENE_COMPONENT_TRANSFORM)
-  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_TRANSFORM, {
-    defaultData: SCENE_COMPONENT_TRANSFORM_DEFAULT_VALUES,
-    deserialize: deserializeTransform,
-    serialize: serializeTransform
-  })
+  Engine.instance.sceneLoadingRegistry.set(SCENE_COMPONENT_TRANSFORM, {})
 
   const _frustum = new Frustum()
   const _projScreenMatrix = new Matrix4()
